chore(tests): move OverlayFs path-validation tests inline
`tests/integration/overlay_path_validation_tests.rs` exercised
`OverlayFs` directly through the `FileSystem` trait — no `Bash`,
no interpreter, no end-to-end pipeline. It was a unit test of an
internal type living in the wrong place.

Move it next to `OverlayFs` as `fs::overlay::tests::path_validation`.
The block is kept as a nested submodule so the existing flat list of
overlay tests stays readable.

This is the only file in `tests/integration/` that truly fits the
"mis-filed integration test" pattern; the rest legitimately exercise
the public `Bash` API end-to-end and belong where they are.
