+++ crates/bashkit/tests/integration/main.rs
@@ -63,7 +63,6 @@ pub mod mkfifo_tests;
 pub mod nested_subscript_tests;
 pub mod network_security_tests;
 pub mod output_truncation_tests;
-pub mod overlay_path_validation_tests;
 pub mod proptest_differential;
 pub mod python_integration_tests;
 pub mod python_security_tests;
